src/third_party/blink/web_tests/svg/custom/empty-viewbox-attr-expected.txt

CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
